Add timecode for Quote 135 [#151]
This is probably the kind of thing that could be semi-automated at stub
creation time. The "NOTE(annotator)" I made at the start of the stream
really helped to find this quickly in the IRC log, so it seems pretty
doable and shouldn't incur too much of a delay. The only difference here
from my usual process is that I got the timecode from the progress bar
in the video embedded in the website rather than mpv's progress bar, so
there is the potential for the time to be slightly off. Obviously this
couldn't be fully automated because !addquote can only happen after the
quote has been said. Let's see how it pans out.