Add stub for Day 196 [#208]
@HandmadeCompanion/annotation-pushers: For this one I tried out an idea
given to me by @powerc9000, which essentially involved me dropping
NOTE(annotator) into the chat at points where the annotations probably
ought to go. I've left the results of this in the .html.md, commented
out so they hopefully won't go live on the site, for you guys to see.

Naturally, as I was doing it live, the times wouldn't be spot on, and
there was at least one point where I wasn't sure what Casey had done,
but I knew that we'd want an annotation to be left there. So we'll need
to tweak the times - in fact, I've left them in UTC, but ideally we'd
want the script or hmh_bot's command(s) to give us the time relative to
the Start time. Also, the notes really refer to the End Of Topic (EOT),
so they'll have to be offset by one, if that makes sense.

What do you lot think of it? May this kind of thing be helpful, or is it
just confusing? If we think it could help us, then we could see about
giving hmh_bot a command or two just for the Annotation Pushers to use,
which will leave these notes for us.