Annotate Day 196 [#208]
@HandmadeCompanion/annotation-pushers

So, the live annotation notes did seem to help after all, and weren't
too distracting. I was a little afraid that I may rely too heavily on
them and overlook some things needing timecodes that I'd missed during
the stream, but this didn't really happen. Because I annotated it the
usual way, watching the whole video from start to end (Q&A), I could
just add a timecode if one was missing.

I'll probably try this again, and reckon someone else in the team ought
to try it, just to see if they find it useful. Essentially, all you need
to do is to watch the stream live and to leave a NOTE(annotator) in the
chat at the end of each topic.

Also, as we're nearing the end of the backlog, we'll need to think about
organising ourselves as to who does what once we are up to date. I'm
thinking it'd be nice to keep the team as flexible as possible, maybe
with some people willing to annotate regularly and others annotating
less regularly who would let us know a day or so before they would like
to do an annotation. This way we could keep the Episode Guide up to date
every day, and give the whole team the opportunity to contribute. What
do you guys think?